requirements.txt
# Core ML Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=1.7.0,<3.0.0
lightgbm>=4.0.0,<5.0.0
imbalanced-learn>=0.11.0,<1.0.0

# Data Validation & Processing
pydantic>=2.0.0,<3.0.0

# Experiment Tracking & Model Registry
mlflow>=2.7.0,<3.0.0
optuna>=3.3.0,<4.0.0

# API Framework
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# Data Versioning
dvc>=3.0.0,<4.0.0
dvc-s3>=3.0.0,<4.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-mock>=3.11.0,<4.0.0

# Code Quality
black>=23.0.0,<25.0.0
flake8>=6.0.0,<8.0.0
isort>=5.12.0,<6.0.0
pre-commit>=3.4.0,<4.0.0

# Monitoring & Logging
prometheus-client>=0.17.0,<1.0.0
evidently>=0.4.0,<1.0.0
python-json-logger>=2.0.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
joblib>=1.3.0,<2.0.0
tqdm>=4.66.0,<5.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.17.0,<6.0.0
shap>=0.43.0,<1.0.0
